+++ src/kandel/kandelSeeder.ts
@@ -10,15 +10,24 @@ import KandelDistribution from "./kandelDistribution";
 import KandelConfiguration from "./kandelConfiguration";
 import { NewKandelEvent } from "../types/typechain/KandelSeeder";
 import { NewAaveKandelEvent } from "../types/typechain/AaveKandelSeeder";
+import { NewSmartKandelEvent } from "../types/typechain/SmartKandelSeeder";
 import GeometricKandelInstance from "./geometricKandel/geometricKandelInstance";
+import SmartKandelInstance from "./smartKandelInstance";
+
+/**
+ * The type of Kandel to create, either a SmartKandel, AaveKandel, or a standard Kandel.
+ */
+export type KandelType = "aave" | "smart" | "simple";
 
 /** The parameters for sowing the Kandel instance.
- * @param onAave Whether to create an AaveKandel which supplies liquidity on Aave to earn yield, or a standard Kandel.
+ * @param type The type of Kandel to create, either a SmartKandel, AaveKandel, or a standard Kandel.
  * @param market The market to create the Kandel for.
  * @param liquiditySharing Whether to enable liquidity sharing for the Kandel so that the signer can publish the same liquidity for multiple router-based Kandels (currently AaveKandel).
  */
-export type KandelSeed = {
-  onAave: boolean;
+export type KandelSeed<
+  TKandelType extends KandelType | undefined = KandelType | undefined,
+> = {
+  type?: TKandelType;
   market: Market;
   liquiditySharing: boolean;
 };
@@ -30,6 +39,7 @@ class KandelSeeder {
   tradeEventManagement: TradeEventManagement = new TradeEventManagement();
 
   aaveKandelSeeder?: typechain.AaveKandelSeeder;
+  smartKandelSeeder?: typechain.SmartKandelSeeder;
   kandelSeeder: typechain.KandelSeeder;
 
   /** Constructor
@@ -56,44 +66,72 @@ class KandelSeeder {
         this.mgv.signer,
       );
     } catch (e) {}
+    try {
+      const smartKandelSeederAddress = Mangrove.getAddress(
+        "SmartKandelSeeder",
+        this.mgv.network.name,
+      );
+      this.smartKandelSeeder = typechain.SmartKandelSeeder__factory.connect(
+        smartKandelSeederAddress,
+        this.mgv.signer,
+      );
+    } catch (e) {}
   }
 
   /** Create a new Kandel instance.
    * @param seed The parameters for sowing the Kandel instance.
    */
-  public async sow(
-    seed: KandelSeed,
+  public async sow<TKandelType extends KandelType | undefined = undefined>(
+    seed: KandelSeed<TKandelType>,
     overrides: ethers.Overrides = {},
-  ): Promise<Transaction<GeometricKandelInstance>> {
-    if (seed.liquiditySharing && !seed.onAave) {
+  ): Promise<
+    Transaction<
+      TKandelType extends "smart"
+        ? SmartKandelInstance
+        : GeometricKandelInstance
+    >
+  > {
+    if (seed.liquiditySharing && seed.type !== "aave") {
       throw Error(
         "Liquidity sharing is only supported for AaveKandel instances.",
       );
     }
 
-    if (seed.onAave && !this.aaveKandelSeeder) {
+    if (seed.type === "aave" && !this.aaveKandelSeeder) {
       throw Error("AaveKandelSeeder is not available on this network.");
     }
 
-    const response = seed.onAave
-      ? this.aaveKandelSeeder!.sow(
-          seed.market.olKeyBaseQuote,
-          seed.liquiditySharing,
-          overrides,
-        )
-      : this.kandelSeeder.sow(
-          seed.market.olKeyBaseQuote,
-          seed.liquiditySharing,
-          overrides,
-        );
+    if (seed.type === "smart" && !this.smartKandelSeeder) {
+      throw Error("SmartKandelSeeder is not available on this network.");
+    }
+
+    const response =
+      seed.type === "aave"
+        ? this.aaveKandelSeeder!.sow(
+            seed.market.olKeyBaseQuote,
+            seed.liquiditySharing,
+            overrides,
+          )
+        : seed.type === "smart"
+          ? this.smartKandelSeeder!.sow(
+              seed.market.olKeyBaseQuote,
+              seed.liquiditySharing,
+              overrides,
+            )
+          : this.kandelSeeder.sow(
+              seed.market.olKeyBaseQuote,
+              seed.liquiditySharing,
+              overrides,
+            );
 
     const func = async (
       response: Promise<ethers.ethers.ContractTransaction>,
     ) => {
       const receipt = await (await response).wait();
-      return await this.getKandelFromReceipt({
+
+      return await this.getKandelFromReceipt<TKandelType>({
         receipt,
-        onAave: seed.onAave,
+        type: seed.type ?? ("simple" as any),
         market: seed.market,
       });
     };
@@ -104,21 +142,32 @@ class KandelSeeder {
   /** Gets the Kandel instance created in a transaction via sow.
    * @param params The parameters.
    * @param params.receipt The receipt of the transaction.
-   * @param params.onAave Whether the Kandel is an AaveKandel.
+   * @param params.type The type of Kandel created, either a SmartKandel, AaveKandel, or a standard Kandel.
    * @param params.market The market the Kandel is for.
    * @returns The Kandel instance created in the transaction.
    */
-  public async getKandelFromReceipt(params: {
+  public async getKandelFromReceipt<
+    TKandelType extends KandelType | undefined = undefined,
+  >(params: {
     receipt: ethers.ethers.ContractReceipt;
-    onAave: boolean;
+    type: TKandelType;
     market: Market;
-  }) {
-    if (params.onAave && !this.aaveKandelSeeder) {
+  }): Promise<
+    TKandelType extends "smart" ? SmartKandelInstance : GeometricKandelInstance
+  > {
+    if (params.type === "aave" && !this.aaveKandelSeeder) {
       throw Error("AaveKandelSeeder is not available on this network.");
+    }
+    if (params.type === "smart" && !this.smartKandelSeeder) {
+      throw Error("SmartKandelSeeder is not available on this network.");
     }
     const events = this.tradeEventManagement.getContractEventsFromReceipt(
       params.receipt,
-      params.onAave ? this.aaveKandelSeeder! : this.kandelSeeder,
+      params.type === "aave"
+        ? this.aaveKandelSeeder!
+        : params.type === "smart"
+          ? this.smartKandelSeeder!
+          : this.kandelSeeder,
     );
     for (const evt of events) {
       const name = "event" in evt ? evt.event : "name" in evt ? evt.name : null;
@@ -127,6 +176,15 @@ class KandelSeeder {
           const kandelEvent = evt as NewKandelEvent;
           return GeometricKandelInstance.create({
             address: kandelEvent.args.kandel,
+            signer: this.mgv.signer,
+            market: params.market,
+          }) as any;
+        }
+        case "NewSmartKandel": {
+          evt as NewSmartKandelEvent;
+          const smartKandelEvent = evt as NewSmartKandelEvent;
+          return SmartKandelInstance.create({
+            address: smartKandelEvent.args.kandel,
             signer: this.mgv.signer,
             market: params.market,
           });
@@ -138,7 +196,7 @@ class KandelSeeder {
             address: aaveKandelEvent.args.aaveKandel,
             signer: this.mgv.signer,
             market: params.market,
-          });
+          }) as any;
         }
       }
     }
@@ -148,17 +206,22 @@ class KandelSeeder {
   }
 
   /** Retrieves the default gasreq for the Kandel type.
-   * @param onAave Whether to get the gasreq for an AaveKandel or a standard Kandel.
+   * @param type The type of Kandel to get the default gasreq for. If null, then the default gasreq for the standard Kandel will be used.
    * @returns The gasreq for the Kandel type.
    */
-  public async getDefaultGasreq(onAave: boolean) {
-    if (onAave && !this.aaveKandelSeeder) {
+  public async getDefaultGasreq(type?: KandelType) {
+    if (type === "aave" && !this.aaveKandelSeeder) {
       throw Error("AaveKandelSeeder is not available on this network.");
     }
+    if (type === "smart" && !this.smartKandelSeeder) {
+      throw Error("SmartKandelSeeder is not available on this network.");
+    }
     return (
-      onAave
+      type === "aave"
         ? await this.aaveKandelSeeder!.KANDEL_GASREQ()
-        : await this.kandelSeeder.KANDEL_GASREQ()
+        : type === "smart"
+          ? await this.smartKandelSeeder!.KANDEL_GASREQ()
+          : await this.kandelSeeder.KANDEL_GASREQ()
     ).toNumber();
   }
 
@@ -194,7 +257,7 @@ class KandelSeeder {
           this.configuration.getConfig(seed.market).gaspriceFactor,
         gasprice,
       ),
-      gasreq: gasreq ?? (await this.getDefaultGasreq(seed.onAave)),
+      gasreq: gasreq ?? (await this.getDefaultGasreq(seed.type)),
     });
   }
 
@@ -202,17 +265,17 @@ class KandelSeeder {
    * @param params The parameters.
    * @param params.market The market the Kandel is deployed to.
    * @param params.offerType The type of offer.
-   * @param params.onAave Whether the Kandel is an AaveKandel.
+   * @param params.type The type of Kandel to get the minimum volume for. If null, then the standard Kandel will be used.
    * @param params.factor The factor to multiply the minimum volume by. Defaults to minimumBasePerOfferFactory / minimumQuotePerOfferFactor from KandelConfiguration.
    * @returns The minimum recommended volume.
    */
   public async getMinimumVolume(params: {
     market: Market;
     offerType: Market.BA;
-    onAave: boolean;
+    type?: KandelType;
     factor?: number;
   }) {
-    const gasreq = await this.getDefaultGasreq(params.onAave);
+    const gasreq = await this.getDefaultGasreq(params.type);
     return this.getMinimumVolumeForGasreq({ ...params, gasreq });
   }