+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Next version
 
+- Add Smart kandels to SDK
+- Changed `onAave` params to `type` params
+
 # 2.0.5-43
 
 - Move `Transaction` type from `Market` to `util/transactions`. Also, use this type in `MangroveAmplifier` instead of the duplicated type in that file.